refactor: moving evm handlers to vm folder
