+++ sgxvm/src/utils.rs
@@ -3,6 +3,14 @@ use crate::ocall;
 use crate::types::{Allocation, AllocationWithResult};
 use sgx_types::sgx_status_t;
 
+use crate::vm::utils::convert_topic_to_proto;
+use crate::protobuf_generated::ffi::HandleTransactionResponse;
+use crate::vm::types::ExecutionResult;
+use crate::protobuf_generated::ffi::Topic;
+use crate::protobuf_generated::ffi::Log;
+
+use protobuf::Message;
+
 /// Allocates provided data outside of enclave
 /// * data - bytes to allocate outside of enclave
 /// 
@@ -30,4 +38,43 @@ pub fn allocate_inner(data: Vec<u8>) -> AllocationWithResult {
             AllocationWithResult::default()
         }
     }
+}
+
+/// Converts raw execution result into protobuf and returns it outside of enclave
+pub fn convert_and_allocate_transaction_result(
+    execution_result: ExecutionResult,
+) -> AllocationWithResult {
+    let mut response = HandleTransactionResponse::new();
+    response.set_gas_used(execution_result.gas_used);
+    response.set_vm_error(execution_result.vm_error);
+    response.set_ret(execution_result.data);
+
+    // Convert logs into proper format
+    let converted_logs = execution_result
+        .logs
+        .into_iter()
+        .map(|log| {
+            let mut proto_log = Log::new();
+            proto_log.set_address(log.address.as_fixed_bytes().to_vec());
+            proto_log.set_data(log.data);
+
+            let converted_topics: Vec<Topic> =
+                log.topics.into_iter().map(convert_topic_to_proto).collect();
+            proto_log.set_topics(converted_topics.into());
+
+            proto_log
+        })
+        .collect();
+
+    response.set_logs(converted_logs);
+
+    let encoded_response = match response.write_to_bytes() {
+        Ok(res) => res,
+        Err(err) => {
+            println!("Cannot encode protobuf result. Reason: {:?}", err);
+            return AllocationWithResult::default();
+        }
+    };
+
+    allocate_inner(encoded_response)
 }
